index_main.cpp now does abstract_index/index.cpp
